Please sonarlint, make list model provider a function type
